pyproject.toml
[project]
name = "ag"
version = "0.1.0"
description = "AI-powered cil tool"
authors = [{name = "HeZzz", email = "hez2z@foxmail.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "pyautogui>=0.9.53",
    "click>=8.1.0",
    "python-dotenv>=1.0.1",
    "pillow>=10.4.0",
]

[project.scripts]
ag = "main:main"

[tool.poetry]
name = "ag"
version = "0.1.0"
description = "AI-powered cil tool"
authors = ["HeZzz hez2z@foxmail.com"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
openai = "^1.0.0"
pyautogui = "^0.9.53"
click = "^8.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
